Move response_handler to transaction
