+++ lib/connection/transaction.ex
@@ -70,4 +70,44 @@ defmodule Janus.Connection.Transaction do
         true
     end
   end
+
+  @spec handle_transaction({:ok, any} | {:error, any, any}, binary, atom | :ets.tid()) :: :ok
+  def handle_transaction(response, transaction, pending_calls_table) do
+    transaction_status = transaction_status(pending_calls_table, transaction)
+
+    case transaction_status do
+      {:ok, from} ->
+        GenServer.reply(from, response)
+        :ets.delete(pending_calls_table, transaction)
+
+      {:error, :outdated} ->
+        :ets.delete(pending_calls_table, transaction)
+
+      _ ->
+        # NOOP
+        nil
+    end
+
+    call_result =
+      case response do
+        {:ok, _} -> "OK"
+        {:error, _, _} -> "ERROR"
+      end
+
+    log_transaction_status(transaction_status, transaction, response, call_result)
+  end
+
+  defp log_transaction_status({:ok, _from}, transaction, data, call_result) do
+    "[#{__MODULE__} #{inspect(self())}] Call #{call_result}: transaction = #{inspect(transaction)}, data = #{
+      inspect(data)
+    }"
+    |> Logger.debug()
+  end
+
+  defp log_transaction_status({:error, reason}, transaction, data, call_result) do
+    "[#{__MODULE__} #{inspect(self())}] Received #{call_result} reply to the #{reason} call: transaction = #{
+      inspect(transaction)
+    }, data = #{inspect(data)}"
+    |> Logger.warn()
+  end
 end
